requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
discord.py==1.2.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
requests==2.22.0
soupsieve==1.9.5
typing-extensions==3.7.4.1
urllib3==1.25.6
websockets==6.0
yarl==1.3.0
